Adding a separate class file for dqn.
